Refactorizar entidades al nivel de repositorios
