+++ Backend API/src/Core/index.js
@@ -1,9 +1,7 @@
-import Entities from "./Abstractions/Entities/index";
 import Enums from "./Abstractions/Enums/index";
 import Interfaces from "./Abstractions/Interfaces";
 
 module.exports = {
-    Entities,
     Enums,
     Interfaces
 };
