+++ Backend API/src/Repositories/Entities/employee.class.js
@@ -1,4 +1,4 @@
-import ICreatedModifiedOn from "../Interfaces/i-created-modified-on.interface";
+import ICreatedModifiedOn from "../../Core/Abstractions/Interfaces/i-created-modified-on.interface";
 
 class Employee extends ICreatedModifiedOn {
     id; // PK
